People memory: one concern per module for the scribe and the node glue
scribe.py kept seven jobs; it now keeps the worker, and the conversation
as the scribe reads it (transcript), the prompt and response schema
(scribe_prompt), the answer as data (scribe_output), the rule set that
gates a proposal into writes (scribe_rules), the on-disk hour of windows
(scribe_queue) and deep recall (recall) each live in their own module.
node_adapters.py kept ~500 lines of pure functions beside the ROS glue;
the parameters as plain data (node_config), the left-eye stream and its
duty cycle (camera_feed), what may be said about the tracks (publishing)
and the service bookkeeping (mutations) are their own modules. Every
definition has exactly one home; no re-export shims. Behaviour unchanged.
